added click limiters on buttons in admin and staff fragments

+++ app/src/main/java/cvsu/clearance/app/AdminViewStationFragment.java
@@ -7,6 +7,7 @@ import android.content.Intent;
 import android.net.Uri;
 import android.os.Bundle;
 import android.os.Handler;
+import android.os.SystemClock;
 import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -75,6 +76,7 @@ public class AdminViewStationFragment extends Fragment implements AdapterView.On
     String StationRequirementCatch;
     String StationLocationCatch;
     String StationIsRequiredCatch;
+    private long mLastClickTime = 0;
 
     boolean signaturedoc,signaturefile;
 
@@ -165,6 +167,13 @@ public class AdminViewStationFragment extends Fragment implements AdapterView.On
         updateButton.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
+                // This method prevents user from clicking the button too much.
+                // It only last for 1.5 seconds.
+                if (SystemClock.elapsedRealtime() - mLastClickTime < 1500){
+                    return;
+                }
+                mLastClickTime = SystemClock.elapsedRealtime();
+
                 performValidation();
                 uploadFile();
             }
@@ -173,6 +182,13 @@ public class AdminViewStationFragment extends Fragment implements AdapterView.On
         deleteButton.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
+                // This method prevents user from clicking the button too much.
+                // It only last for 1.5 seconds.
+                if (SystemClock.elapsedRealtime() - mLastClickTime < 1500){
+                    return;
+                }
+                mLastClickTime = SystemClock.elapsedRealtime();
+
                 deleteStation();
             }
         });
@@ -180,6 +196,13 @@ public class AdminViewStationFragment extends Fragment implements AdapterView.On
         fileButton.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
+                // This method prevents user from clicking the button too much.
+                // It only last for 1.5 seconds.
+                if (SystemClock.elapsedRealtime() - mLastClickTime < 1500){
+                    return;
+                }
+                mLastClickTime = SystemClock.elapsedRealtime();
+
                 openFileChooser();
             }
         });
